# docs/FE-contract/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "HistoryTalk API",
    "description": "API Specification for HistoryTalk project, synced with FE and latest DB specifications.",
    "version": "2.1"
  },
  "servers": [
    {
      "url": "http://localhost:8080/api/v1",
      "description": "Development Server"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/staff/quizzes": {
      "get": {
        "tags": ["Quiz - Manage"],
        "summary": "Lấy danh sách bộ Quiz cho Admin (Phân trang)",
        "parameters": [
          { "name": "search", "in": "query", "schema": { "type": "string" } },
          { "name": "era", "in": "query", "schema": { "$ref": "#/components/schemas/Era" } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 0 } },
          { "name": "size", "in": "query", "schema": { "type": "integer", "default": 10 } }
        ],
        "responses": {
          "200": {
            "description": "Thành công",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ResponseWrapper" },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "allOf": [
                            { "$ref": "#/components/schemas/Pagination" },
                            {
                              "type": "object",
                              "properties": {
                                "content": {
                                  "type": "array",
                                  "items": { "$ref": "#/components/schemas/ContentAdminQuizSet" }
                                }
                              }
                            }
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Quiz - Manage"],
        "summary": "Tạo bộ Quiz mới",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string" },
                  "contextId": { "type": "string" },
                  "era": { "$ref": "#/components/schemas/Era" },
                  "questions": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "content": { "type": "string" },
                        "options": { "type": "array", "items": { "type": "string" } },
                        "correctAnswer": { "type": "integer" },
                        "explanation": { "type": "string" }
                      },
                      "required": ["content", "options", "correctAnswer"]
                    }
                  }
                },
                "required": ["title", "contextId", "era"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tạo thành công",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ResponseWrapper" },
                    {
                      "type": "object",
                      "properties": {
                        "data": { "$ref": "#/components/schemas/ContentAdminQuizSet" }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/staff/quizzes/{quizId}": {
      "get": {
        "tags": ["Quiz - Manage"],
        "summary": "Chi tiết bộ Quiz Admin",
        "parameters": [
          { "name": "quizId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Thành công",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ResponseWrapper" },
                    {
                      "type": "object",
                      "properties": {
                        "data": { "$ref": "#/components/schemas/ContentAdminQuizSet" }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Quiz - Manage"],
        "summary": "Cập nhật thông tin bộ Quiz",
        "parameters": [
          { "name": "quizId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string" },
                  "contextId": { "type": "string" },
                  "era": { "$ref": "#/components/schemas/Era" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Thành công",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ResponseWrapper" },
                    {
                      "type": "object",
                      "properties": {
                        "data": { "$ref": "#/components/schemas/ContentAdminQuizSet" }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Quiz - Manage"],
        "summary": "Xóa vĩnh viễn Quiz",
        "parameters": [
          { "name": "quizId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Thành công"
          }
        }
      }
    },
...
    "/staff/quizzes/{quizId}/questions": {
      "post": {
        "tags": ["Quiz - Manage"],
        "summary": "Thêm câu hỏi mới vào bộ Quiz",
        "parameters": [
          { "name": "quizId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "content": { "type": "string" },
                  "options": { "type": "array", "items": { "type": "string" } },
                  "correctAnswer": { "type": "integer" },
                  "explanation": { "type": "string" }
                },
                "required": ["content", "options", "correctAnswer"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Thêm câu hỏi thành công",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ResponseWrapper" },
                    {
                      "type": "object",
                      "properties": {
                        "data": { "$ref": "#/components/schemas/QuizQuestion" }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/staff/quizzes/{quizId}/questions/{questionId}": {
      "put": {
        "tags": ["Quiz - Manage"],
        "summary": "Sửa câu hỏi trong bộ Quiz",
        "parameters": [
          { "name": "quizId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "questionId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "content": { "type": "string" },
                  "options": { "type": "array", "items": { "type": "string" } },
                  "correctAnswer": { "type": "integer" },
                  "explanation": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sửa thành công"
          }
        }
      },
      "delete": {
        "tags": ["Quiz - Manage"],
        "summary": "Xóa câu hỏi khỏi bộ Quiz",
        "parameters": [
          { "name": "quizId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "questionId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Xóa thành công"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Era": {
        "type": "string",
        "enum": ["ANCIENT", "MEDIEVAL", "MODERN", "CONTEMPORARY"]
      },
      "ResponseWrapper": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "message": { "type": "string" },
          "data": { "type": "object" },
          "timestamp": { "type": "string", "format": "date-time" }
        },
        "required": ["success", "message", "timestamp"]
      },
      "Pagination": {
        "type": "object",
        "properties": {
          "totalElements": { "type": "integer" },
          "totalPages": { "type": "integer" },
          "currentPage": { "type": "integer" },
          "pageSize": { "type": "integer" },
          "hasNext": { "type": "boolean" },
          "hasPrevious": { "type": "boolean" }
        },
        "required": ["totalElements", "totalPages", "currentPage", "pageSize", "hasNext", "hasPrevious"]
      },
...
      "QuizQuestion": {
        "type": "object",
        "properties": {
          "questionId": { "type": "string" },
          "content": { "type": "string" },
          "options": {
            "type": "array",
            "items": { "type": "string" },
            "minItems": 4,
            "maxItems": 4
          },
          "correctAnswer": { "type": "integer" },
          "explanation": { "type": "string" }
        },
        "required": ["questionId", "content", "options", "correctAnswer"]
      },
      "ContentAdminQuizSet": {
        "type": "object",
        "properties": {
          "quizId": { "type": "string" },
          "title": { "type": "string" },
          "era": { "$ref": "#/components/schemas/Era" },
          "playCount": { "type": "integer" },
          "rating": { "type": "number" },
          "contextId": { "type": "string" },
          "contextTitle": { "type": "string" },
          "createdBy": { "type": "string" },
          "createdDate": { "type": "string", "format": "date-time" },
          "updatedDate": { "type": "string", "format": "date-time" },
          "isActive": { "type": "boolean" },
          "questions": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/QuizQuestion" }
          }
        },
        "required": ["quizId", "title", "era", "playCount", "rating", "contextId", "contextTitle", "createdBy", "createdDate", "updatedDate", "isActive", "questions"]
      }
    }
  }
}